chore(playwright): capture settings screenshots across modes
